backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic[email]>=2.10
python-jose[cryptography]>=3.3
pymongo[srv]>=4.6
slowapi>=0.1.9
python-multipart>=0.0.18
httpx>=0.28
pydantic-settings>=2.7
